requirements.txt
adbdgl-adapter==2.1.0
python-arango==7.5.7